+++ src/app/shared/main/main.component.ts
@@ -1,16 +1,10 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  input,
-  signal,
-} from '@angular/core';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 
 import { SummaryComponent } from './components/summary/summary.component';
 import { HabilitiesComponent } from './components/habilities/habilities.component';
 import { TrainingComponent } from './components/training/training.component';
 import { LanguagesComponent } from './components/languages/languages.component';
 import { WorkExperiencesComponent } from './components/work-experience/work-experiences.component';
-import { defaultLang } from '../../interfaces/langs';
 
 @Component({
   selector: 'app-main',